Extract pure chat-cursor geometry predicates into their own file
model_pty_status.go (461 lines) mixed two separable concerns: pure
chat-cursor geometry predicates on snapshots (isChatInputCursorPosition,
isRenderableChatCursorPosition, isStoredChatCursorPosition,
isBottomEdgeCornerPosition, isSuspiciousBottomEdgeCornerCursor — no *Model
receiver) and Model agent-activity/state queries. Move the five pure
package-level predicates into a new model_pty_status_cursor_geometry.go
(carrying only the compositor import); model_pty_status.go drops to 369
lines. Pure mechanical move, zero behavior change.
